Automatically stop "left over" Nessie processes
Use Gradle build service mechanism to automatically stop all Nessie
Quarkus processes at the end of each build.

Background is that if a Gradle task failed, the remaining task
actions may not be executed, leading to a "lost" Nessie Quarkus
process.

The fix here is to use a Gradle build-service, which is guaranteed
to be "closed" once the build finishes.

Side note: there is already a Java shutdown hook to kill potentially
left-over Nessie Quarkus processes, but since the Gradle daemon is
kept alive, the Nessie Quarkus processes will not be cleaned up via
that mechanism.